Fixes source mapping in Firefox.
I made this change locally per recommendation in #4 and it seems to resolve the problem.  I also tested in Chrome, which appeared to work for me.

Breakpoints in Firefox still don't work from within source maps, but I think that is a Firefox bug, not an issue with this library.

I recommend having either CI or another human manually test this change.  I have no idea what I'm doing and just took the code referenced in #4 and pasted it verbatim.

Fixes #4